+++ test-project/src/helpers.cljs
@@ -1,0 +1,4 @@
+(ns helpers)
+
+(defn greet [name]
+  (str "Hello, " name "!"))