+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 [Squint](https://github.com/squint-cljs/squint): Light-weight ClojureScript dialect
 
+## Unreleased
+
+- Fix [#783](https://github.com/squint-cljs/squint/issues/783): qualified symbol resolution in macro expansions
+
 ## v0.10.186 (2026-03-07)
 
 - Fix #799: forgot to add `squint.math` file to `package.json`